simplified code to calculate entropy
